backend/requirements.txt
# Kairos Backend Dependencies
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
aiofiles
google-generativeai
python-dotenv
httpx
pydantic
